Moved relevant entity view stuff into model.
